Refactor solvers
